secure-file-storage/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pycryptodome
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-dotenv
sqlalchemy